initial Coming Soon placeholder
Just a background image of madison with the words Coming Soon
centered on the screen.

Not using any sort of UI framework yet, so this was more
annoying than I am used to. I miss MUI, but I think I may plug in
Chakra if I get the change since people seem to like it. I didn't
do it yet because I don't want to futz with getting emotion to work
in SSR and on the Client (yet).