feat: write a cookie when adding an address
We move our add form from "/" to "/wi/madison" so we can
support more cities. Our index is empty and in the next commit
we can list our addresses or link to the new address page for
Madison.

I've written an AddressStore class that abstracts away a cookie
so we can easily support functions like add, remove, reorder, clear
for our addresses. Currently add is supported.